# app/docs/api/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "GenAI Demo - DDD 電商平台 API",
    "description" : "基於領域驅動設計（DDD）的電商平台 API\n\n## 架構特色\n- 遵循 DDD 分層架構\n- 六邊形架構（Hexagonal Architecture）\n- CQRS 和事件驅動設計\n- 豐富的領域模型\n\n## 主要功能模組\n- 訂單管理：完整的訂單生命週期\n- 支付處理：多種支付方式和狀態管理\n- 庫存管理：即時庫存追蹤和預留\n- 產品管理：產品資訊和分類\n- 促銷活動：多樣化的促銷規則\n- 定價策略：動態定價和佣金計算\n- 客戶管理：會員等級和偏好設定\n\n## 錯誤處理\n所有 API 都遵循統一的錯誤回應格式，包含錯誤代碼、訊息、時間戳記和詳細資訊。\n",
    "contact" : {
      "name" : "GenAI Demo Team",
      "url" : "https://github.com/humank/genai-demo",
      "email" : "contact@genai-demo.com"
    },
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://test-api.genai-demo.com",
    "description" : "測試環境"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/stats" : {
      "get" : {
        "tags" : [ "統計報表" ],
        "summary" : "獲取系統統計數據",
        "description" : "獲取系統整體統計數據，包括各個數據表的記錄數量、系統狀態等基礎統計信息",
        "operationId" : "getStats",
        "responses" : {
          "200" : {
            "description" : "成功獲取統計數據",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                },
                "examples" : {
                  "統計數據範例" : {
                    "description" : "統計數據範例",
                    "value" : {
                      "customers" : 150,
                      "orders" : 1250,
                      "products" : 85,
                      "payments" : 1180,
                      "inventory_items" : 320,
                      "status" : "success",
                      "message" : "統計數據獲取成功"
                    }
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "系統內部錯誤",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/stats/payment-methods" : {
      "get" : {
        "tags" : [ "統計報表" ],
        "summary" : "獲取支付方式分布統計",
        "description" : "獲取系統中所有支付交易按支付方式分布的統計數據，用於分析客戶支付偏好和支付渠道效果",
        "operationId" : "getPaymentMethodStats",
        "responses" : {
          "200" : {
            "description" : "成功獲取支付方式統計",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                },
                "examples" : {
                  "支付方式統計範例" : {
                    "description" : "支付方式統計範例",
                    "value" : {
                      "paymentMethodDistribution" : {
                        "CREDIT_CARD" : 450,
                        "DEBIT_CARD" : 280,
                        "BANK_TRANSFER" : 150,
                        "DIGITAL_WALLET" : 320,
                        "CASH_ON_DELIVERY" : 95
                      },
                      "status" : "success",
                      "message" : "支付方式統計獲取成功"
                    }
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "系統內部錯誤",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/stats/order-status" : {
      "get" : {
        "tags" : [ "統計報表" ],
        "summary" : "獲取訂單狀態分布統計",
        "description" : "獲取系統中所有訂單按狀態分布的統計數據，用於分析訂單處理情況和業務狀態",
        "operationId" : "getOrderStatusStats",
        "responses" : {
          "200" : {
            "description" : "成功獲取訂單狀態統計",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                },
                "examples" : {
                  "訂單狀態統計範例" : {
                    "description" : "訂單狀態統計範例",
                    "value" : {
                      "statusDistribution" : {
                        "PENDING" : 45,
                        "CONFIRMED" : 120,
                        "PROCESSING" : 85,
                        "SHIPPED" : 200,
                        "DELIVERED" : 350,
                        "CANCELLED" : 25
                      },
                      "status" : "success",
                      "message" : "訂單狀態統計獲取成功"
                    }
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "系統內部錯誤",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "StandardErrorResponse" : {
        "required" : [ "code", "message", "path", "timestamp" ],
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string",
            "description" : "錯誤代碼",
            "example" : "VALIDATION_ERROR"
          },
          "message" : {
            "type" : "string",
            "description" : "錯誤訊息",
            "example" : "請求參數驗證失敗"
          },
          "timestamp" : {
            "type" : "string",
            "description" : "錯誤發生時間",
            "format" : "date-time"
          },
          "path" : {
            "type" : "string",
            "description" : "請求路徑",
            "example" : "/api/orders"
          },
          "details" : {
            "type" : "array",
            "description" : "詳細錯誤資訊",
            "nullable" : true,
            "items" : {
              "$ref" : "#/components/schemas/FieldError"
            }
          },
          "traceId" : {
            "type" : "string",
            "description" : "追蹤ID，用於問題排查",
            "nullable" : true,
            "example" : "abc123-def456-ghi789"
          }
        },
        "description" : "標準錯誤回應"
      },
      "FieldError" : {
        "required" : [ "field", "message" ],
        "type" : "object",
        "properties" : {
          "field" : {
            "type" : "string",
            "description" : "欄位名稱",
            "example" : "customerId"
          },
          "message" : {
            "type" : "string",
            "description" : "錯誤訊息",
            "example" : "客戶ID不能為空"
          },
          "rejectedValue" : {
            "type" : "object",
            "description" : "拒絕的值",
            "nullable" : true,
            "example" : null
          },
          "code" : {
            "type" : "string",
            "description" : "錯誤代碼",
            "nullable" : true,
            "example" : "NotBlank"
          }
        },
        "description" : "欄位錯誤詳情",
        "nullable" : true
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "請在此處輸入 JWT Token，格式：Bearer {token}",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "apiKey" : {
        "type" : "apiKey",
        "description" : "請在此處輸入 API Key",
        "name" : "X-API-Key",
        "in" : "header"
      }
    }
  }
}